requirements.txt
geopy==2.2.0
folium==0.12.1
requests==2.26.0
Flask==2.1.0
Werkzeug==2.0.2
pyYaml==6.0.1
